Unit testing refactor
