+++ main/CMakeLists.txt
@@ -3,4 +3,5 @@ idf_component_register(SRCS
 "kernel/kernel.cpp" "kernel/phases.cpp" "kernel/sensors.cpp"
 "controller/open_loop.cpp"
 "time_series/filters.cpp"
+"testing/test_phases.cpp"
                     INCLUDE_DIRS ".")
